Renamed "Ollama Chat" to "Ollama Generate" and added new "Ollama Chat" node using /api/chat
