+++ src/index.ts
@@ -1,5 +1,6 @@
 import type { RivetPlugin, RivetPluginInitializer } from "@ironclad/rivet-core";
-import { ollamaChat } from "./nodes/OllamaChatNode";
+import { ollamaChat } from "./nodes/OllamaGenerateNode";
+import { ollamaChat2 } from "./nodes/OllamaChatNode";
 import { getOllamaModel } from "./nodes/GetOllamaModelNode";
 import { listOllamaModels } from "./nodes/ListOllamaModelsNode";
 import { pullModelToOllama } from "./nodes/PullModelToOllamaNode";
@@ -30,6 +31,7 @@ const plugin: RivetPluginInitializer = (rivet) => {
 
     register: (register) => {
       register(ollamaChat(rivet));
+      register(ollamaChat2(rivet));
       register(getOllamaModel(rivet));
       register(listOllamaModels(rivet));
       register(pullModelToOllama(rivet));
